airflow/dags/Scripts/Pipeline_Scripts/PyPDF/requirements.txt
boto3
python-dotenv
PyPDF2